Fix: no need for CurrentProtocol pattern, just use Providing protocols
