--- OpenWeather/App/CurrentWeather/CurrentLocationProviding.swift
@@ -1,27 +1,0 @@
-//
-//  OpenWeather template generated by OpenBytes on 15/03/2023.
-//
-// Created by Ahmed Shendy.
-//  CurrentLocationProviding.swift
-//
-
-import Foundation
-import CoreLocation
-
-protocol CurrentLocationProviding {
-    mutating func locationName(for location: CLLocation) async throws -> String
-}
-
-struct CurrentLocationProvider: CurrentLocationProviding, LocationProviding {
-    func locationName(for location: CLLocation) async throws -> String {
-        let response: LocationNameResponse = try await locationName(for: location)
-
-        return response.name
-    }
-}
-
-struct MockCurrentLocationProvider: CurrentLocationProviding, WeatherProviding {
-    func locationName(for location: CLLocation) async throws -> String {
-        return "location-name"
-    }
-}
