--- OpenWeather/App/CurrentWeather/CurrentWeatherViewModel.swift
@@ -1,29 +1,0 @@
-//
-//  OpenWeather template generated by OpenBytes on 15/03/2023.
-//
-// Created by Ahmed Shendy.
-//  CurrentWeatherViewModel.swift
-//
-
-import Foundation
-
-final class CurrentWeatherViewModel: ObservableObject, ErrorHandling {
-    private var currentWeatherProviding: CurrentWeatherProviding
-    private var currentLocationProviding: CurrentLocationProviding
-
-    var errorHandler: ErrorHandler
-
-    init(
-        currentWeatherProviding: CurrentWeatherProviding,
-        currentLocationProviding: CurrentLocationProviding,
-        errorHandler: ErrorHandler = ErrorHandler(
-            plugins: [
-                ToastErrorPlugin()
-            ]
-        )
-    ) {
-        self.currentWeatherProviding = currentWeatherProviding
-        self.currentLocationProviding = currentLocationProviding
-        self.errorHandler = errorHandler
-    }
-}
