+++ OpenWeather/App/CurrentWeather/WeatherDetailsViewModel.swift
@@ -1,0 +1,29 @@
+//
+//  OpenWeather template generated by OpenBytes on 15/03/2023.
+//
+// Created by Ahmed Shendy.
+//  WeatherDetailsViewModel.swift
+//
+
+import Foundation
+
+final class WeatherDetailsViewModel: ObservableObject, ErrorHandling {
+    private var weatherProviding: WeatherProviding
+    private var locationProviding: LocationProviding
+
+    var errorHandler: ErrorHandler
+
+    init(
+        weatherProviding: WeatherProviding,
+        locationProviding: LocationProviding,
+        errorHandler: ErrorHandler = ErrorHandler(
+            plugins: [
+                ToastErrorPlugin()
+            ]
+        )
+    ) {
+        self.weatherProviding = weatherProviding
+        self.locationProviding = locationProviding
+        self.errorHandler = errorHandler
+    }
+}
