+++ OpenWeather/Networking/Implementation/WeatherProvider.swift
@@ -2,27 +2,29 @@
 //  OpenWeather template generated by OpenBytes on 15/03/2023.
 //
 // Created by Ahmed Shendy.
-//  CurrentWeatherProvider.swift
+//  WeatherProvider.swift
 //
 
 import Foundation
+import WeatherKit
 import CoreLocation
 
-protocol CurrentWeatherProviding {
-    mutating func currentWeather(
-        for location: CLLocation
-    ) async throws -> DeviceWeather
-}
+struct WeatherProvider: WeatherProviding {
+    private let weatherService: WeatherService = .shared
 
-struct CurrentWeatherProvider: CurrentWeatherProviding, WeatherProviding {
     func currentWeather(for location: CLLocation) async throws -> DeviceWeather {
         let response = try await weather(for: location)
 
         return try WeatherAdapter.device(from: response.weather)
     }
+
+    private func weather(for location: CLLocation) async throws -> WeatherResponse {
+        let weather = try await weatherService.weather(for: location)
+        return WeatherResponse(weather: weather)
+    }
 }
 
-struct MockCurrentWeatherProvider: CurrentWeatherProviding, WeatherProviding {
+struct MockWeatherProvider: WeatherProviding {
     func currentWeather(for location: CLLocation) async throws -> DeviceWeather {
         return DeviceWeather(
             currentTemperature: .init(value: 0, unit: .celsius),
