--- OpenWeather/Networking/Implementation/WeatherProviding+Default.swift
@@ -1,19 +1,0 @@
-//
-//  OpenWeather template generated by OpenBytes on 15/03/2023.
-//
-// Created by Ahmed Shendy.
-//  WeatherProviding+Default.swift
-//
-
-import Foundation
-import WeatherKit
-import CoreLocation
-
-extension WeatherProviding {
-    var weatherService: WeatherService { .shared }
-
-    func weather(for location: CLLocation) async throws -> WeatherResponse {
-        let weather = try await weatherService.weather(for: location)
-        return WeatherResponse(weather: weather)
-    }
-}
