+++ OpenWeather/Networking/Interface/LocationProviding.swift
@@ -9,5 +9,5 @@ import Foundation
 import CoreLocation
 
 protocol LocationProviding {
-    func locationName(for location: CLLocation) async throws -> LocationNameResponse
+    func locationName(for location: CLLocation) async throws -> String
 }
